refactor(usage): give each breakdown sub-component its own folder
Row, List and ScopeTotal were flat files in components/. Each now has a
folder and a barrel, per the component structure rule. Row in particular
had its own stylesheet sitting beside it.

Names the tri-state a scope reports while it is doing so: undefined is
still in flight, null is answered with nothing, a number is answered.
That was written out three times. UsageScope moves to utils alongside it
so the components take their types from one place rather than from each
other.
