+++ frontend/web/components/pages/usage/components/UsageBreakdown/components/List/index.ts
@@ -1,0 +1,2 @@
+export { default } from './List'
+export type { ListProps } from './List'
