+++ frontend/web/components/pages/usage/components/UsageBreakdown/components/Row/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Row'
+export type { RowProps } from './Row'
