+++ frontend/web/components/pages/usage/components/UsageBreakdown/components/ScopeTotal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ScopeTotal'
