+++ frontend/web/components/pages/usage/components/UsageBreakdown/index.ts
@@ -1,5 +1,5 @@
 export { default } from './UsageBreakdown'
 export { default as ScopeTotal } from './components/ScopeTotal'
 export type { UsageBreakdownProps } from './UsageBreakdown'
-export type { BreakdownDimension, BreakdownRow } from './utils'
+export type { BreakdownDimension, BreakdownRow, UsageScope } from './utils'
 export { useUsageBreakdown } from './useUsageBreakdown'
