+++ frontend/web/components/pages/usage/components/UsageBreakdown/useScopedBreakdown.ts
@@ -1,6 +1,5 @@
 import { useCallback, useEffect, useMemo, useState } from 'react'
-import { fromScopedTotals } from './utils'
-import { UsageScope } from './components/ScopeTotal'
+import { fromScopedTotals, ScopedTotal, UsageScope } from './utils'
 
 export const useScopedBreakdown = (scopes: UsageScope[], identity: string) => {
   const [totals, setTotals] = useState<Record<string, number | null>>({})
@@ -9,17 +8,14 @@ export const useScopedBreakdown = (scopes: UsageScope[], identity: string) => {
     setTotals({})
   }, [identity])
 
-  const onTotal = useCallback(
-    (key: string, total: number | null | undefined) => {
-      if (total === undefined) {
-        return
-      }
-      setTotals((current) =>
-        current[key] === total ? current : { ...current, [key]: total },
-      )
-    },
-    [],
-  )
+  const onTotal = useCallback((key: string, total: ScopedTotal) => {
+    if (total === undefined) {
+      return
+    }
+    setTotals((current) =>
+      current[key] === total ? current : { ...current, [key]: total },
+    )
+  }, [])
 
   const rows = useMemo(() => fromScopedTotals(scopes, totals), [scopes, totals])
